requirements.txt
PySide2>=5.15.2
pytube>=10.4.1
ffmpeg-python>=0.2.0
git+git://github.com/nficano/pytube.git#egg=pytube